Model the provider webpage off the AWS provider documentation (#17)
Rewrite the documentation of the provider to follow that of the AWS Provider. This will cover usage, and outline how to provide the sufficient credentials to the provider.

At the time there is no method for generating these kinds of markdown files, so this has been written up manually. You can see the reference for existing websites in the terraform-provider-aws repository.